+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Changed
 - Moved translations to a submodule for Crowdin support
+- Merged `[p]progress` and `[p]profile lookup` commands
 
 ## [2020.30] - 2020-07-25
 ### Added